Synchronize access to the marshaller map.
I'm pretty sure that concurrent modification of the marshaller map is causing my super weird missing
marshaller problem. In the last instance of the bug, I had a map of marshallers that had some of the
primitive marshallers, but missing others, which makes no sense outside of something like
concurrency.